improve(ui): a better hydration story for the header/status
Moving all static parts of the header into the Layout itself saves us
from having to remember to include the header in all pages. A separate
Status component can be inserted next to the header logo if a page wants.
This also means the status is the only thing that needs hydration, which
doesn't affect performance, but the self-contained nature of it feels
nicer.